refactor: extract common section headline into component
