+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2019-02-08T09:48:36.050Z\n"
-"PO-Revision-Date: 2019-02-08T09:48:36.050Z\n"
+"POT-Creation-Date: 2019-02-08T10:12:46.569Z\n"
+"PO-Revision-Date: 2019-02-08T10:12:46.569Z\n"
 
 msgid "Cancel"
 msgstr ""
